extracting fragment into a separate file
